+++ src/next/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+0.1.2
+++++++
+* Fix the bug that the nested command history could not be logged
+
+0.1.1
+++++++
+* Support personalized recommendation
+
 0.1.0
 ++++++
 * Initial release.